+++ main.go
@@ -19,12 +19,13 @@ import (
 	"os"
 
 	// +kubebuilder:scaffold:imports
+	argocdV1alpha1 "github.com/argoproj/argo-cd/v2/pkg/apis/application/v1alpha1"
 	certmanagerV1 "github.com/jetstack/cert-manager/pkg/apis/certmanager/v1"
 	servicecatalogv1beta1 "github.com/kubernetes-sigs/service-catalog/pkg/apis/servicecatalog/v1beta1"
 	claimV1alpha1 "github.com/tmax-cloud/hypercloud-multi-operator/apis/claim/v1alpha1"
 	clusterV1alpha1 "github.com/tmax-cloud/hypercloud-multi-operator/apis/cluster/v1alpha1"
-	claimcontroller "github.com/tmax-cloud/hypercloud-multi-operator/controllers/claim"
-	clustercontroller "github.com/tmax-cloud/hypercloud-multi-operator/controllers/cluster"
+	claimController "github.com/tmax-cloud/hypercloud-multi-operator/controllers/claim"
+	clusterController "github.com/tmax-cloud/hypercloud-multi-operator/controllers/cluster"
 	k8scontroller "github.com/tmax-cloud/hypercloud-multi-operator/controllers/k8s"
 	traefikV1alpha1 "github.com/traefik/traefik/v2/pkg/provider/kubernetes/crd/traefik/v1alpha1"
 
@@ -53,6 +54,7 @@ func init() {
 	utilruntime.Must(servicecatalogv1beta1.AddToScheme(scheme))
 	utilruntime.Must(certmanagerV1.AddToScheme((scheme)))
 	utilruntime.Must(traefikV1alpha1.AddToScheme(scheme))
+	utilruntime.Must(argocdV1alpha1.AddToScheme(scheme))
 	// +kubebuilder:scaffold:scheme
 }
 
@@ -84,7 +86,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	if err = (&claimcontroller.ClusterClaimReconciler{
+	if err = (&claimController.ClusterClaimReconciler{
 		Client: mgr.GetClient(),
 		Log:    ctrl.Log.WithName("controllers").WithName("ClusterClaim"),
 		Scheme: mgr.GetScheme(),
@@ -92,7 +94,7 @@ func main() {
 		setupLog.Error(err, "unable to create controller", "controller", "ClusterClaim")
 		os.Exit(1)
 	}
-	if err = (&clustercontroller.ClusterManagerReconciler{
+	if err = (&clusterController.ClusterManagerReconciler{
 		Client: mgr.GetClient(),
 		Log:    ctrl.Log.WithName("controllers").WithName("ClusterManager"),
 		Scheme: mgr.GetScheme(),
@@ -133,7 +135,7 @@ func main() {
 	// 	setupLog.Error(err, "unable to create controller", "controller", "serviceController")
 	// 	os.Exit(1)
 	// }
-	if err = (&clustercontroller.ClusterRegistrationReconciler{
+	if err = (&clusterController.ClusterRegistrationReconciler{
 		Client: mgr.GetClient(),
 		Log:    ctrl.Log.WithName("controllers").WithName("ClusterRegistration"),
 		Scheme: mgr.GetScheme(),