+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # HyperCloud-Multi-Operator changelog!!
 All notable changes to this project will be documented in this file.
 
+<!-------------------- v5.0.31.0 start -------------------->
+
+## HyperCloud-Multi-Operator_5.0.31.0 (2022. 08. 12. (금) 10:38:29 KST)
+
+### Added
+
+### Changed
+
+### Fixed
+
+### CRD yaml
+
+### Etc
+
+<!--------------------- v5.0.31.0 end --------------------->
+
 <!-------------------- v5.0.30.0 start -------------------->
 
 ## HyperCloud-Multi-Operator_5.0.30.0 (2022. 08. 02. (화) 13:09:21 KST)
@@ -122,7 +138,7 @@ All notable changes to this project will be documented in this file.
 ## HyperCloud-Multi-Operator_5.0.26.14 (2022. 07. 25. (월) 09:38:00 KST)
 
 ### Added
-
+  - [feat] log level설정할 수 있도록 코드 추가 및 yaml 수정 by seung 
 ### Changed
   - [mod] traefik service가 single cluster에 아직 설치되지 않았을 경우, error가 아닌 info로 로그 레벨 변경 by SISILIA
   - [mod] lb type의 svc를 지우는 로직에서 error 반환 로직 수정 by SISILIA